requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
reportlab>=3.6.0
tkinter>=8.6.0
joblib>=1.1.0